Where's my coffee? - Fix for #2341
When a context is first initialized (cold start) and used the Entity Framework loads and validates the model, and the result is cached at the AppDomain level. This is a one-time cost that allows users to get good performance when they instantiate context instances. It is expected that after the first time the model is loaded, the cost of instantiating a context instance (warm start) is constant and low.

Contexts that derive from DbContext have been known to have a higher cost for warm start than those derived from ObjectContext because of the additional structrures that DbContext contains. One of these structures is the dictionary of CLR type to Entity Set which is being rebuilt every time a context gets initialized, despite having the exact same result every time and not much of a memory impact. For small models rebuilding this dictionary represents a very low cost but programs that have large models and make use of the context-per-request pattern will see the thoughput affected negatively.

The change here is to cache the CLR type to Entity Set dictionary at the metadata workspace level for the following reasons:
  1. The mapping is immutable once the model is loaded.
  2. The mapping is derived from the O-space and C-space data stored in the metadata workspace.
  3. Allows the cache to be disposed of when the metadata workspace is disposed.
  4. The cache only consumes a few bytes even for very large models.

As an example: a very large test model went from having a warm throughput of 16 RPS to 810 RPS. The performance difference will vary depending on the scenario.