Being original... (Make sure initializer runs only once even when not using PersistSecurityInfo)
We keep a record of whether or not an initializer has been run for a given connection. This is used to ensure that the initializer is only run once. The connection string is used as part of the key for this lookup. If the connection string is changes then the initializer will run again for the new conncetion. This is by design.

However, sometimes the connection string can be manipulated silently, such as when a username and password is supplied but PersistSecurityInfo is not used. In this case the username and password are removed after the connection has been opened. This can cause the initializer to run again for what is, from the app's perspective, the same connection string.

The fix is to use the OriginalConnectionString for the lookup key. This is the connection string provided to the context before any operation is performed with it.