Second cup of coffee - Fix for 2341
After submitting a fix for #2341 with changeset 35835b580fb we detected a threading issue with the entity set mappings cache. The fix is to localize the lock and the cache at the metadata workspace level. This also looks cleaner on the InternalContext side since there's no real need to keep a reference to the mapping dictionary there.

The access to the metadata workspace through the ObjectContext has also been studied in order to make sure we're not initializing the ObjectContext when it's not needed. It looks like the change is safe.

All tests are passing.