[Polycheck violation; title suppressed] (2306: Suppress finalizers in InterceptableDbCommand)
InterceptableDbCommand ultimately inherits from Component which has a finalizer, but DbCommands are not disposed by EF and InterceptableDbCommand doesn't suppress finalizers. This means that InterceptableDbCommands and all the objects referenced from them do not get garbage collected until after the finalizer thread gets a chance to finalize them. This in turn can result in out-of-memory when the finalizer thread can't keep up.

The initial fix is to make sure finalizers are suppressed. See the issue for discussion of further changes that might be made.