Playing checkers in public... (Make DatabaseTableChecker functionality public so that other providers can implement)
This change makes public the DatabaseTableChecker functionality that is currently only implemented for SQL Server and SQL Server CE. There is no change to the way the functionality is implemented and no default implementation for other providers. The idea is simply to allow other providers to do what SQL Server and SQL CE currently do, if this is useful for other providers.

The changes are basically:
- Make public the interface implemented by providers and make it a base class
- Resolve an instance of this class from DbConfiguration
- Move the SQL Server and SQL CE implementations into their respective provider assemblies
- Make public some methods around transactions and interception so that these methods can be used from provider assemblies
- Have the SQL Server and SQL CE providers register the service in the normal way