Hybrid technology requires more gas... (2254: TPT/TPH hybrid generates database, but derived properties are not nullable)
The default configuration is TPH and columns coming from properties in derived types are made nullable. When mapping switches to TPT the nullability of these columns is reset. But if the mapping is actually hybrid with TPT at the base and then TPH below it then the nullability of some columns should not be reset.

Note that there is a fairly good chance that this change will cause a perf hit for model building with large models containing inheritance. We should investigate/profile as necessary.