Hey Mom, that's my stuff! (2418: ObjectContext.ExecuteStoreQueryInternal disposes DbCommand too early)
In previous versions of EF the command was not disposed at all. See https://entityframework.codeplex.com/workitem/2306 for more info. In EF6 we started disposing the commands immediately, which is wrong because the command should not be disposed until after the reader is disposed. The easiest fix would be to revert to the EF5 behavior. In this case I chose to instead keep disposing but only when the returned ObjectResult is disposed. This seems like the least change in EF6 functionality that fixes the issue.