Schrödinger's Granny (2435: Stop IsPropertyChanged from throwing for change tracking entities)
The IsPropertyChanged API was added in .NET 4.5. The intention of this API is to check whether a property's current value is different from its original value. This may be different from the actual modified state of the property for properties that have been explicitly set as modified or where the property was changed but then changed back.

The problem is that for change tracking proxies and entities that derive from EntityObject there is no way of knowing this information because original values are not tracked for this type of entity. Therefore the best we can do is return whether or not the property is marked as modified, which is what we now do. We could throw instead, but for most usages the modified state is sufficient and hence seems like the more useful behavior.