Bug 952621: DTS: EF creates unnecessary joins when using derived types.
When the PreProcessor encounters a navigation PropertyOp it rewrites the corresponding node into one that represents a query. The query is then added as nested subquery to the first RelOp ancestor. When the RelOp nodes are processed, the SubqueryTrackingVisitor creates a chain of outer applies: if S1, S2, ... are the subqueries and D the original (driver) input, the output created looks like OuterApply(OuterApply(D, S1), S2), ... Later in the plan compiler, when the transformations are run, the outer applies are transformed into left outer joins. If a Linq query projects multiple properties of an object accessed via a navigation property (x.NavProp.P1, x.NavProp.P2, ...) the final query ends having duplicated joins. The JoinElimination phase tries to eliminated duplicates but does not cover complex cases.

Instead of attempting to enhance the JoinElimination phase further, and make that code even more complex, the fix eliminates the duplicate subqueries early, in the PreProcessor, by caching and reusing the rewritten nodes. The logic for cache hits is implemented in the NavigationPropertyOpInfo class. Basically two navigation property nodes are considered equivalent from the cache point of view if they have the same RelOp ancestor (not null), same navigation property (EdmMember) and the subtrees having them as root are equivalent. For fast retrieval GetHashCode computes a hash code using the RelOp ancestor's hash code, the navigation property's hash code and the node's computed hash value.