Racing flags for Marty McFly - fix for #2445 The EF designer exposes a performance regression in Sql Server 2014 when an EDMX model is updated from the database. The regression was introduced in Sql Server 2014 by a change in their algorithm for cardinality estimation. Enabling flag 9481 changes the algorithm back to what it was in previous version of the product. This only affects the connection used by the EF designer.
Note that we're only enabling this flag for connections to Sql Server, determined by means of the provider invariant name. Given that this is a Sql Server flag, enabling against older versions of Sql Server won't have an effect nor will it return any error or warning. If a specified trace flag is not one that affects a query execution plan, the option will be silently ignored.