Foot in the door... (2447: Need a public method in DbModelBuilder)
This change creates a public non-generic method for registering entity types on DbModelBuilder. It does not allow further configuration of the entity type since this would require quite extensive changes to the fluent API surface which is heavily based on generics. However, these is still some value to this method because it is not uncommon to want to register types dynamically and then have the model configured by conventions--for example, this is what DbContext does by default. EF7 will allow full configuration without using generics.

I gave this method a different name than the current internal method since I know a fair amount of code already accesses that method using Reflection. The return type of the public method is void and both making the internal method public and changing its return type could cause Reflection code to break.