Fix for Issue 2267: Code First, table splitting: Sequence contains more than one matching element
Issue:
When configuring the principal key column names we assumed that the principal entity type from the pair sharing the table is configured first.

Solution:
Configure the dependent key columns to match the principal ones even if the dependent entity type was configured first.
Also removed the KeyNamesType annotation as it doesn't appear to serve any purpose.