Bug 952621: Second issue with previous fix for duplicate joins (duplicate key exception while adding to dictionary)
For the dictionary lookup and additions to function properly the key's hash code has to be deterministic, and the fields involved in the Equals comparison have to be immutable. Apparently there are scenarios when _node is changed by the Rewrite or Visit calls that happen between TryGetValue and Add, in the VisitNavPropertyOp method. The fix seals a NavigationPropertyOpInfo instance by cloning the _node field. Thus, changes to the original node will not affect the result of Equals. Cloning _node does not clone the referenced navigation property (same instance is used). Also the roots are compared by reference so changes in their sub-trees do not affect the result of Equals.

While this is theoretically likely the reason for the failure, the customer did not provide a repro and I wasn't able to come up with one so far.