CodePlex 2365: DropTable DB-Hang
DROP TABLE executed outside a transaction hangs if ALTER TABLE has been executed within the transaction. In the migrations' Down methods we generate DropForeignKey operations (i.e. ALTER TABLE) before DropTable. In lots of cases this is not necessary and the issue can be worked-around by removing the DropForeignKey operations and arranging the DROP table operations to drop first those that reference other tables to be dropped. However this does not work if circular foreign key constraints exist. For that reason this change groups subsequent statements with SuppressTransaction = false and executes them within separate transactions that are committed before the statements with SuppressTransaction = false are executed. With this change the migration statements can potentially execute in separate transactions, thus not ensuring full rollback in case of failure. However those cases should be rare, if any, because in the majority of the cases the statements with SuppressTransaction = true are (or can be placed) all at the beginning or end.