Cosmo and Dibs...  (Fix for 2382: Explicit Composite Index Only Includes One of Two Specified Columns)
The issue here was the foreign key index convention which was replacing any existing index on the column with the convention index. This didn't usually happen because if an explicit index is defined for the FK columns then the convention doesn't kick in. In this case the explicit index only spanned one column of the composite foreign key and so the convention still kicked in and the existing annotation on the column was clobbered.