Fix for issue #2557: Query: Make GetMethod used in QueryableExtensions compatible with Roslyn C# compiler
Problem:
The GetMethod helper method used in QueryableExtension to initialize static fields with MethodInfos for the aggregate LINQ operators has a very rigid dependency on the internal implementation details of lambdas and that is causing the problems since those details have changed in Roslyn.

Solution:
Use DynamicInvoke to handle all implementations of the lambdas being used.