Fix to CodePlex #2595 - VerifyTypeSupportedForComparison
Issue was that when using ODATA expand with DTO projections in some cases. Basically, expand adds a check for null for on the referenced entity. However we the entities were projected as DTOs, the null check would be performed on the DTO, and not the entity. Previously we would allow IsNull checks for complex types and row types, by recursively checking all their properties/fields as well. We did not accept a collection as a valid type to compare - which could cause an exception.

Fix is to loosen the validation and allow arbitraty row to be IsNull compared (still throwing for equal comparison for same cases as before). We still have a problem of this check always returning false (DTO is never null, instead all it's fields are null) - this however is consistent with our previous behavior for less complex cases. CR: Diego