Fix for #139: PowerTools: setting |datadirectory| path fails if using non-default output path
Setting an output path that not a child of the directory containing the project will result in a target directory that contains path traversals, e.g. C:\EfTestProg\EfTestProg..\Output\Debug\ but SqlClient isn't happy with those. Fix is to add a call to Path.GetFullPath() when computing the data directory.
Verification: smoke testing.