Fix to #6 - .NET/EF6 ObjectContext leak using compiled LINQ query that references String.StartsWith()
Problem was that when translating StartsWith/EndsWith/Contains into Like, we are escaping the argument in the runtime using ProviderManifest. Originally we were doing this by adding Func rooted in ExpressionConverter as part of the QueryParameterExpression. ExpressionConverter in turn holds reference to the context which caused the original problem (since the parameter is being added to the cache, and hence persists through the garbage collection)

Fix is to sever the connection to ExpressionConverter by using Expression<Func<>> instead and pass ProviderManifest a constant, rather than closure. Also, original method contained out parameter, which now have been converted to a Tuple<> result.