Fix for #165: Add flag to control whether scope_identity() is used to retrieve generated values for integer columns
Currently the logic is hardcoded in the DmlSqlGenerator class of the SQL
Server provider for EF, in particular in the
IsValidScopeIdentityColumnType() method to use scope_identity() to
recover generated values for any columns for which it may apply.

This PR adds a flag UseScopeIdentity with a default value of true in
SqlProviderServices (in the past we added TruncateDecimalsToScale in a
similar fashion). If the flag is set to false, the following SQL will be
generated which uses the OUTPUT clause:

``` SQL
DECLARE @generated_keys table([Id] int)
INSERT [dbo].[People]([Name])
OUTPUT inserted.[Id] INTO @generated_keys
VALUES (@0)
SELECT t.[Id]
FROM @generated_keys AS g JOIN [dbo].[People] AS t ON g.[Id] = t.[Id]
WHERE @@ROWCOUNT > 0
```

TO-DO:
- run all existing tests with the flag changed and verify anything that
breaks
- add a targed test
- do some perf verification
- if this causes a significant perf regression we can consider a fix
that allows disabling scope_identity on specific columns