Cream filling to no-ads @lastwishes.us
Title is just a "template" or "named-topic"...etc.  Our vision is limited by the company we keep and Organizing tribunal trifection infection of seeking perfection. Leadership is abusiveness of power instead of action of a individuals forward movement.  Productive individuals are  under the microscope while the strict views others around  these power houses wayward thinkers are compressed, oppressed, and obsessed over.  Me,Myself, and I have enough, just knowing the uncompromising attributes of under the radar / slips in the cracks /  peer-pressure suicides / lime light life/ or whatever it was before is slang today is lying, & dying in vain really beyond my belief that humanity can and will overcome the game of blaming and just offer logical solutions to be a catalyst of elements bring the perfect storm,or tsunami tidal wave, to burden their faults so others may have safe passage on a globalization scale. That Be A Sweet Victory For All!!!